rename `tauri` module to `primitives`
