+++ core/tauri/src/path/init.js
@@ -2,11 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-Object.defineProperty(window.__TAURI__.__INTERNALS__, 'path', {
+Object.defineProperty(window.__TAURI__.__INTERNALS__.plugins, 'path', {
   value: {
-    path: {
-      sep: __TEMPLATE_sep__,
-      delimiter: __TEMPLATE_delimiter__
-    }
+    sep: __TEMPLATE_sep__,
+    delimiter: __TEMPLATE_delimiter__
   }
 })
